Auto stash before merge of "master" and "origin/master"

+++ bar/platoListado.php
@@ -15,10 +15,11 @@ $sql = "
                     p.nombre AS pNombre,
                     p.precio as pPrecio,
                     c.id     AS cId,
+                    p.estrella as pEstrella,
                     c.nombre AS cNombre
                 FROM
                    plato AS p INNER JOIN categoria AS c
-                   ON p.categoriaId = c.id
+                   ON p.categoria_id = c.id
                 ORDER BY p.nombre
         ";
 
@@ -49,22 +50,37 @@ $rs = $select->fetchAll();
 <table border='1'>
 
     <tr>
+        <th>Estrella</th>
         <th>Nombre</th>
         <th>Categoria</th>
     </tr>
 
     <?php foreach ($rs as $fila) {?>
-        <tr>
-
+         <tr>
+             <?php
+        if($estrella){
+            if($fila["pEstrella"] == "1"){?>
                 <td><a href='platoFicha.php?id=<?=$fila["pId"]?>'> <?=$fila["pNombre"]?> </a></td>
                 <td><a href=  'categoriaFicha.php?id=<?=$fila["cId"]?>'> <?=$fila["cNombre"] ?></a></td>
                 <td><a href='platoEliminar.php?id=<?=$fila["pId"]?>'> (X)                   </a></td>
-                <?
-            ?>
-
-        </tr>
+                <?php
+                    }
+                } else{
+            if($fila["pEstrella"] == "1"){?>
+                <td><a href='platoEstablecerEstadoEstrella.php?id=<?=$fila["pId"]?>'><img src="/bar/img/estrellaRellena.png"  width="20" height="20"></a></td>
+                <?php
+            }else{?>
+                <td><a href='platoEstablecerEstadoEstrella.php?id=<?=$fila["pId"]?>'><img src="/bar/img/estrellaVacia.png"  width="20" height="20"></a></td>
+                <?php
+            }?>
+        <td><a href='platoFicha.php?id=<?=$fila["pId"]?>'> <?=$fila["pNombre"]?> </a></td>
+        <td><a href=  'categoriaFicha.php?id=<?=$fila["cId"]?>'> <?=$fila["cNombre"] ?></a></td>
+        <td><a href='platoEliminar.php?id=<?=$fila["pId"]?>'> (X)                   </a></td>
+    <?php
+    }
+    ?>
     <?php } ?>
-
+         </tr>
 </table>
 
 <br />
